:recycle: rename file name and directory name
